case 25684 - check for file existance before delete

+++ NbtLogic/csw/Schema/CswNbtSchemaModTrnsctn.cs
@@ -1428,8 +1428,14 @@ namespace ChemSW.Nbt.Schema
             }
             finally
             {
-                File.Delete( BatchFilePath );
-                File.Delete( SqlFilePath );
+                if( File.Exists( BatchFilePath ) )
+                {
+                    File.Delete( BatchFilePath );
+                }
+                if( File.Exists( SqlFilePath ) )
+                {
+                    File.Delete( SqlFilePath );
+                }
             }
 
         } // runExternalSqlScript
